Rewrite work with db using cursor and loader. Moved view holder to separate class.

+++ mobile/src/main/java/com/elegion/androidschool/finalproject/event/ListSelectedEvent.java
@@ -4,13 +4,13 @@ package com.elegion.androidschool.finalproject.event;
  * Created by Aleksandra on 17.10.15.
  */
 public class ListSelectedEvent {
-    private int id;
+    private long id;
 
-    public ListSelectedEvent(int id) {
+    public ListSelectedEvent(long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 }
